+++ GRDB/Core/Database.swift
@@ -121,8 +121,7 @@ public final class Database {
                     throw DatabaseError(code: readCode, message: String.fromCString(sqlite3_errmsg(sqliteConnection)))
                 }
             } catch {
-                // deinit is not called: close connection
-                sqlite3_close_v2(sqliteConnection)
+                close()
                 throw error
             }
         #endif
@@ -142,13 +141,39 @@ public final class Database {
         setupDefaultCollations()
     }
     
-    deinit {
+    private var isClosed: Bool = false
+    func close() {
+        preconditionValidQueue()
+        assert(!isClosed)
+        
+        updateStatementCache = [:]
+        selectStatementCache = [:]
+        
+        // sqlite3_close_v2 was added in SQLite 3.7.14 http://www.sqlite.org/changes.html#version_3_7_14
+        // It is available from iOS 8.2 and OS X 10.10 https://github.com/yapstudios/YapDatabase/wiki/SQLite-version-(bundled-with-OS)
         configuration.SQLiteConnectionWillClose?(sqliteConnection)
-        let code = sqlite3_close_v2(sqliteConnection)
-        guard code == SQLITE_OK else {
-            fatalError(DatabaseError(code: code, message: lastErrorMessage).description)
+        if #available(iOS 8.2, OSX 10.10, *) {
+            let code = sqlite3_close_v2(sqliteConnection)
+            guard code == SQLITE_OK else {
+                fatalError(DatabaseError(code: code, message: lastErrorMessage).description)
+            }
+        } else {
+            let code = sqlite3_close(sqliteConnection)
+            if code != SQLITE_OK {
+                NSLog("%@", "GRDB could not close database with error \(code): \(lastErrorMessage ?? "")")
+                var stmt: SQLiteStatement = sqlite3_next_stmt(sqliteConnection, nil)
+                while stmt != nil {
+                    NSLog("%@", "GRDB unfinalised statement: \(String.fromCString(sqlite3_sql(stmt))!)")
+                    stmt = sqlite3_next_stmt(sqliteConnection, stmt)
+                }
+            }
         }
+        isClosed = true
         configuration.SQLiteConnectionDidClose?()
+    }
+    
+    deinit {
+        assert(isClosed)
     }
     
     func releaseMemory() {
@@ -803,6 +828,13 @@ extension Database {
         // 1   | firstName | TEXT    | 0       | NULL       | 0  |
         // 2   | lastName  | TEXT    | 0       | NULL       | 0  |
         
+        if #available(iOS 8.2, OSX 10.10, *) { } else {
+            // Work around a bug in SQLite where PRAGMA table_info would
+            // return a result even after the table was deleted.
+            if !tableExists(tableName) {
+                throw DatabaseError(message: "no such table: \(tableName)")
+            }
+        }
         let columnInfos = ColumnInfo.fetchAll(self, "PRAGMA table_info(\(tableName.quotedDatabaseIdentifier))")
         guard columnInfos.count > 0 else {
             throw DatabaseError(message: "no such table: \(tableName)")