+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 Release Notes
 =============
 
+## 0.63.0
+
+Released May 17, 2016
+
+**Fixed**
+
+- Restored support for iOS before 8.2 and OS X before 10.10. Fixes [#51](https://github.com/groue/GRDB.swift/issues/51).
+
+**Breaking Changes**
+
+- Support for advanced migrations is not available until iOS 8.2 and OS X 10.10:
+    
+    ```diff
+     public struct DatabaseMigrator {
+    -    public mutating func registerMigration(identifier: String, withDisabledForeignKeyChecks disabledForeignKeyChecks: Bool = false, migrate: (Database) throws -> Void)
+    +    public mutating func registerMigration(identifier: String, migrate: (Database) throws -> Void)
+    +    @available(iOS 8.2, OSX 10.10, *)
+    +    public mutating func registerMigrationWithDisabledForeignKeyChecks(identifier: String, migrate: (Database) throws -> Void)
+    ```
+
+
 ## 0.62.0
 
 Released May 12, 2016